requirements.txt
faiss-cpu
sentence-transformers
langchain
transformers
peft
fastapi
uvicorn
requests
sympy
sqlite3